requirements.txt
fastapi
uvicorn
python-dotenv
openai
chromadb
pypdf
python-multipart
sentence-transformers
